pyproject.toml
[project]
name = "nfldata"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "matplotlib==3.10.6",
    "nfl-data-py==0.3.2",
    "pandas-stubs==2.3.2.250827",
    "pyside6==6.9.2",
]
